Negation increases the complexity of a selector by 2 instead of one now.
I changed the way relationship-selectors are built from a perceived relationship. Instead of using the blank selector for selecting the base-object of the relationship, a selector that selects single objects only and selects the base-object is used. Only if there is no such selector, the blank selector is used.

Done: Take the posteriori into account when spreading attention from hypotheses to the features or objects.

I lowered the probability of picking a feature first. Now in 75% of the time, the algorithm picks an object first and perceives an unknown feature on it. Only in 25% of the time the algorithm trys to perceive a popular feature on a new object. This should help with finding additional properties of singled out objects, like needed in PBP 26.