Reintroduced unique and all solution types.
Each solution is keeping track of selects-single-left-object and selects-all-left-objects (+right) and converts a failing solution into unique or all type, if appropriate. To make sure we keep a 1-to-1 relationship between solutions, we only allow one solution per selector (but it can dynamically change its mode).

We should check the "L" or "R" hypotheses earlier than the "LR" ones. I lowered the steepness of the complexity penalty function a bit.